split systemops for operating systems and add linux
