+++ client/internal/routemanager/systemops_windows.go
@@ -1,0 +1,4 @@
+//go:build windows
+// +build windows
+
+package routemanager
